requirements.txt
absl-py==2.1.0
arrow==1.3.0
astunparse==1.6.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cachetools==5.5.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==23.2.1
cloudpickle==3.0.0
codecarbon==2.6.0
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
decorator==5.1.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
dm-tree==0.1.8
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
httplib2==0.20.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keras==2.14.0
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
onnx==1.16.2
onnx-tf==1.10.0
onnxruntime==1.19.2
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
pexpect==4.8.0
pillow==10.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.36
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pybind11==2.13.6
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pynvml==11.5.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
pyzfp==0.5.5
questionary==2.0.1
rapidfuzz==3.9.7
requests==2.25.1
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.1
service-identity==18.1.0
shellingham==1.5.4
six==1.16.0
sos==4.4
ssh-import-id==5.11
sympy==1.13.3
systemd-python==234
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-addons==0.23.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.22.1
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.4.1
torchvision==0.19.1
triton==3.0.0
Twisted==22.1.0
typeguard==2.13.3
typer==0.12.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.14.1
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
